server/utils/token.py
import time
import hashlib
from django.core import signing
from datetime import datetime, timedelta

from passlib.context import CryptContext
from jose import JWTError, jwt


ALGORITHM = "HS256"
SECRET_KEY = "BLOG"

async def create_access_token(data: dict, expires_delta: timedelta = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(hours=8, days=31)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

async def decode_access_token(token: str):
    return jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])