server/requirements.txt
-U
python-jose[cryptography]
robyn
loguru
pymysql
passlib
sqlmodel